fix: resolve CI build error by using local SnakeCaseNamingHelper
The Maliev.Aspire.ServiceDefaults namespace was not being resolved correctly in CI when using PackageReference. This change provides a local internal implementation of the naming helper to ensure build success and maintain database model consistency (snake_case).
